Extract query processing into a standalone helper class and ensure it's covered with tests.
